.gitignore

+++ src/main/java/jp/ebacorp/Hiroyuki/Nishi/EBAtechPMS/UserData/UserRepository.java
@@ -2,5 +2,6 @@ package jp.ebacorp.Hiroyuki.Nishi.EBAtechPMS.UserData;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User,Integer> {
+//User(会員)のリポジトリ。DBのテーブルをJavaのインターフェースにしたようなもの
+public interface UserRepository extends CrudRepository<User,Long> {
 }
